Added support for multiple key directories
A new command line argument has been added to pass the key directory to use.
The use case for this is to have the 2fa profile data stored in an alternative
directory in a removable media, so that there is no need to store it in your
laptop when traveling but in another device such as a USB stick.

For example:
- Create key:
./2fa -d .

- Generate OTP:
./2fa

Note that on OTP generation, the script looks by default in the same directory
where the it's located, that is, both the script and the data can be
stored in the removable device. If you prefer to keep just the profile data in
separate directory, then `-d` can also be used on OTP generation as follows:
./2fa -d <dir>